+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning].
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-01-11
+
 ### Added
 
 - Allow users to derive a key from a secret value and the encryption metadata.
@@ -87,7 +89,8 @@ Initial release.
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 [#6]: https://github.com/apyrgio/tindercrypt/issues/6
 
-[Unreleased]: https://github.com/apyrgio/tindercrypt/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/apyrgio/tindercrypt/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/apyrgio/tindercrypt/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/apyrgio/tindercrypt/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/apyrgio/tindercrypt/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/apyrgio/tindercrypt/compare/v0.1.1...v0.2.0